requirements.txt
# Python 3.10 compatible versions

# Web frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
Flask==2.3.3
Flask-CORS==4.0.0

# Database drivers
motor==3.3.2
pymongo==4.6.0

# Blockchain & Web3
web3==6.15.1
eth-account==0.10.0

# IPFS client (optional)
ipfshttpclient==0.8.0a2

# Environment & configuration
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3

# Authentication & security
python-jose==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP requests
requests==2.31.0